backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
torch
torchvision
numpy
python-dotenv
opencv-python